repo_data/pyproject.toml
[project]
name = "repo-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "daft>=0.4.7",
    "dotenv>=0.9.9",
    "gitpython>=3.1.44",
    "pygithub>=2.6.1",
]
